+++ mathics/builtin/drawing/graphics3d.py
@@ -63,6 +63,19 @@ class Style3D(Style):
         return RGBColor(components=(1, 1, 1, 1))
 
 
+class _Graphics3DElement(InstanceableBuiltin):
+    def init(self, graphics, item=None, style=None):
+        if (
+            item is not None
+            and hasattr(item, "has_form")
+            and not item.has_form(self.get_name(), None)
+        ):
+            raise BoxConstructError
+        self.graphics = graphics
+        self.style = style
+        self.is_completely_visible = False  # True for axis elements
+
+
 class Graphics3D(Graphics):
     r"""
     <dl>
@@ -303,12 +316,3 @@ class Cylinder(Builtin):
             evaluation.error("Cylinder", "oddn", positions)
 
         return Expression("Cylinder", positions, radius)
-
-
-class _Graphics3DElement(InstanceableBuiltin):
-    def init(self, graphics, item=None, style=None):
-        if item is not None and not item.has_form(self.get_name(), None):
-            raise BoxConstructError
-        self.graphics = graphics
-        self.style = style
-        self.is_completely_visible = False  # True for axis elements